fix: Fix NPE error when no field.
